+++ src/resources/users/user.service.ts
@@ -1,11 +1,11 @@
-import * as Hapi from '@hapi/hapi';
-import { Request } from "@hapi/hapi";
+import Hapi, { Request } from '@hapi/hapi';
 import Boom from '@hapi/boom';
 import UserMemoryRepository from './user.memory.repository';
 import TaskMemoryRepository from '../tasks/task.memory.repository';
 import { IUserData, ICreatedUserData } from '../helpers/interfaces';
 import { Users } from '../../entity/users.entity';
 import Logger from '../../logger';
+import { verifyToken } from '../helpers/utils';
 
 export default class UserService {
   /**
@@ -16,10 +16,15 @@ export default class UserService {
    */
   public static getAllUsers = async (request: Request, h: Hapi.ResponseToolkit): Promise<Hapi.ResponseObject> | never => {
     try {
+      const token = await verifyToken(request);
+      if (!token) {
+        Logger.logError('clientError', 'getAllUsers', 'Unauthorized', 401);
+        throw Boom.unauthorized();
+      }
       Logger.logRequestInfo('getAllUsers', request, '../../logs/user-logger.json', 200);
       const allUsers = await UserMemoryRepository.getAllUsers();
       const res = allUsers.length ? allUsers.map(Users.toResponse.bind(Users)) : [];
-      return h.response(res).code(200);
+      return h.response(res).header('Authorization', `Bearer ${token}`).code(200);
     }
     catch (error) {
       if (!Boom.isBoom(error)) {
@@ -39,9 +44,14 @@ export default class UserService {
   public static getUserById = async (request: Request, h: Hapi.ResponseToolkit): Promise<Hapi.ResponseObject> | never => {
     const {userId} = request.params;
     try {
+      const token = await verifyToken(request);
+      if (!token) {
+        Logger.logError('clientError', 'getUserById', 'Unauthorized', 401);
+        throw Boom.unauthorized();
+      }
       Logger.logRequestInfo('getUserById', request, '../../logs/user-logger.json', 200);
       const user = await UserMemoryRepository.getUserById(<string>userId)
-      return h.response(Users.toResponse(user)).code(200);
+      return h.response(Users.toResponse(user)).header('Authorization', `Bearer ${token}`).code(200);
     }
     catch (error) {
       if (!Boom.isBoom(error)) {
@@ -60,11 +70,16 @@ export default class UserService {
    */
   public static updateUserById = async (request: Request, h: Hapi.ResponseToolkit): Promise<Hapi.ResponseObject> | never => {
     try {
+      const token = await verifyToken(request);
+      if (!token) {
+        Logger.logError('clientError', 'updateUserById', 'Unauthorized', 401);
+        throw Boom.unauthorized();
+      }
       Logger.logRequestInfo('updateUserById', request, '../../logs/user-logger.json', 200);
       const payload: IUserData = <IUserData>request.payload;
       const {userId} = request.params;
       const updatedUser: IUserData = await UserMemoryRepository.updateUserById(<string>userId, payload);
-      return h.response(Users.toResponse(updatedUser)).code(200);
+      return h.response(Users.toResponse(updatedUser)).header('Authorization', `Bearer ${token}`).code(200);
     }
     catch (error) {
       if (!Boom.isBoom(error)) {
@@ -83,10 +98,15 @@ export default class UserService {
    */
   public static createUser = async (request: Request, h: Hapi.ResponseToolkit): Promise<Hapi.ResponseObject> | never => {
     try {
+      const token = await verifyToken(request);
+      if (!token) {
+        Logger.logError('clientError', 'createUser', 'Unauthorized', 401);
+        throw Boom.unauthorized();
+      }
       Logger.logRequestInfo('createUser', request, '../../logs/user-logger.json', 201);
       const payload: ICreatedUserData = <ICreatedUserData>request.payload;
       const createdUser = await UserMemoryRepository.createUser(payload);
-      return h.response(Users.toResponse(createdUser)).code(201);
+      return h.response(Users.toResponse(createdUser)).header('Authorization', `Bearer ${token}`).code(201);
     }
     catch (error) {
       if (!Boom.isBoom(error)) {
@@ -105,11 +125,16 @@ export default class UserService {
    */
   public static removeUserById = async (request: Request, h: Hapi.ResponseToolkit): Promise<Hapi.ResponseObject> | never => {
     try {
+      const token = await verifyToken(request);
+      if (!token) {
+        Logger.logError('clientError', 'removeUserById', 'Unauthorized', 401);
+        throw Boom.unauthorized();
+      }
       Logger.logRequestInfo('removeUserById', request, '../../logs/user-logger.json', 204);
       const {userId} = request.params;
       await TaskMemoryRepository.updateTaskByUserId(<string>userId);
       await UserMemoryRepository.removeUserById(<string>userId);
-      return h.response('The user has been deleted').code(204);
+      return h.response('The user has been deleted').header('Authorization', `Bearer ${token}`).code(204);
     }
     catch (error) {
       if (!Boom.isBoom(error)) {
